+++ src/main/java/com/timcooki/jnuwiki/domain/member/controller/AdminController.java
@@ -6,7 +6,7 @@ import com.timcooki.jnuwiki.domain.member.DTO.response.admin.NewReadResDTO;
 import com.timcooki.jnuwiki.domain.member.DTO.response.admin.EditListReadResDTO;
 import com.timcooki.jnuwiki.domain.member.DTO.response.admin.EditReadResDTO;
 import com.timcooki.jnuwiki.util.ApiUtils;
-import com.timcooki.jnuwiki.domain.docsRequest.service.DocsRequestService;
+import com.timcooki.jnuwiki.domain.docsRequest.service.DocsRequestReadService;
 import com.timcooki.jnuwiki.domain.member.service.AdminService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -20,34 +20,34 @@ import org.springframework.web.bind.annotation.*;
 @Slf4j
 @RequiredArgsConstructor
 public class AdminController {
-    private final DocsRequestService docsRequestService;
+    private final DocsRequestReadService docsRequestReadService;
     private final AdminService adminService;
 
     // 문서 기본정보 수정 요청 목록 조회
     @GetMapping("/admin/requests/update")
     private ResponseEntity<?> getModifiedRequests(@PageableDefault(size = 10, sort = "createdAt", direction = Sort.Direction.DESC) Pageable pageable) {
-        EditListReadResDTO modifiedRequests = docsRequestService.getModifiedRequestList(pageable);
+        EditListReadResDTO modifiedRequests = docsRequestReadService.getModifiedRequestList(pageable);
         return ResponseEntity.ok().body(ApiUtils.success(modifiedRequests));
     }
 
     // 새 문서 생성 요청 목록 조회
     @GetMapping("/admin/requests/new")
     private Object getCreatedRequests(@PageableDefault(size = 10, sort = "createdAt", direction = Sort.Direction.DESC) Pageable pageable) {
-        NewListReadResDTO createRequests = docsRequestService.getCreatedRequestList(pageable);
+        NewListReadResDTO createRequests = docsRequestReadService.getCreatedRequestList(pageable);
         return ApiUtils.success(createRequests);
     }
 
     // 문서 기본 정보 수정 요청 상세 조회
     @GetMapping("/admin/requests/update/{docs_request_id}")
     private Object getOneModifiedRequest(@PathVariable("docs_request_id") Long docsRequestId) {
-        EditReadResDTO modifiedRequest = docsRequestService.getOneModifiedRequest(docsRequestId);
+        EditReadResDTO modifiedRequest = docsRequestReadService.getOneModifiedRequest(docsRequestId);
         return ApiUtils.success(modifiedRequest);
     }
 
     // 새 문서 신청 요청 상세 조회
     @GetMapping("/admin/requests/new/{docs_request_id}")
     private Object getOneCreatedRequest(@PathVariable("docs_request_id") Long docsRequestId) {
-        NewReadResDTO modifiedRequest = docsRequestService.getOneCreatedRequest(docsRequestId);
+        NewReadResDTO modifiedRequest = docsRequestReadService.getOneCreatedRequest(docsRequestId);
         return ApiUtils.success(modifiedRequest);
     }
