+++ src/test/java/com/timcooki/jnuwiki/domain/docsRequest/DocsRequestControllerTest.java
@@ -1,26 +1,22 @@
 package com.timcooki.jnuwiki.domain.docsRequest;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.timcooki.jnuwiki.domain.docs.controller.DocsController;
 import com.timcooki.jnuwiki.domain.docs.entity.DocsLocation;
 import com.timcooki.jnuwiki.domain.docsRequest.controller.DocsRequestController;
 import com.timcooki.jnuwiki.domain.docsRequest.dto.request.EditWriteReqDTO;
 import com.timcooki.jnuwiki.domain.docsRequest.dto.request.NewWriteReqDTO;
-import com.timcooki.jnuwiki.domain.docsRequest.dto.response.NewWriteResDTO;
 import com.timcooki.jnuwiki.domain.docsRequest.entity.DocsCategory;
 import com.timcooki.jnuwiki.domain.docsRequest.entity.DocsRequestType;
-import com.timcooki.jnuwiki.domain.docsRequest.service.DocsRequestService;
+import com.timcooki.jnuwiki.domain.docsRequest.service.DocsRequestWriteService;
 import com.timcooki.jnuwiki.domain.member.entity.Member;
 import com.timcooki.jnuwiki.domain.member.entity.MemberRole;
 import com.timcooki.jnuwiki.domain.member.service.MemberService;
 import com.timcooki.jnuwiki.domain.security.config.AuthenticationConfig;
 import com.timcooki.jnuwiki.domain.security.config.JwtFilter;
-import com.timcooki.jnuwiki.domain.security.config.MemberDetails;
 import com.timcooki.jnuwiki.domain.security.service.MemberSecurityService;
 import com.timcooki.jnuwiki.util.errors.GlobalExceptionHandler;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import org.mockito.Mockito;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
 import org.springframework.boot.test.mock.mockito.MockBean;
@@ -28,7 +24,6 @@ import org.springframework.context.annotation.Import;
 import org.springframework.data.jpa.mapping.JpaMetamodelMappingContext;
 import org.springframework.http.MediaType;
 import org.springframework.security.test.context.support.WithMockUser;
-import org.springframework.security.test.context.support.WithUserDetails;
 import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.test.web.servlet.ResultActions;
 import org.springframework.test.web.servlet.request.MockMvcRequestBuilders;
@@ -47,7 +42,7 @@ import java.nio.charset.StandardCharsets;
 public class DocsRequestControllerTest {
 
     @MockBean
-    private DocsRequestService docsRequestService;
+    private DocsRequestWriteService writeService;
 
     @MockBean
     private MemberService memberService;
@@ -82,7 +77,7 @@ public class DocsRequestControllerTest {
                 .build();
 
         String requestBody = om.writeValueAsString(editWriteReqDTO);
-        System.out.println("테스트 : " + requestBody);
+        System.out.println("요청전 : " + requestBody);
 
         // when
         ResultActions result = mvc.perform(
@@ -93,7 +88,7 @@ public class DocsRequestControllerTest {
         );
 
         String responseBody = result.andReturn().getResponse().getContentAsString(StandardCharsets.UTF_8);
-        System.out.println("테스트 : " + responseBody);
+        System.out.println("요청후 : " + responseBody);
 
         // then
         result.andExpect(MockMvcResultMatchers.jsonPath("$.success").value("true"));
@@ -104,7 +99,7 @@ public class DocsRequestControllerTest {
     public void add_create_request() throws Exception {
         // given
         NewWriteReqDTO newWriteReqDTO = NewWriteReqDTO.builder()
-                .docsRequestType(DocsRequestType.MODIFIED)
+                .docsRequestType(DocsRequestType.CREATED)
                 .docsRequestName("ds")
                 .docsRequestLocation(new DocsLocation(12.0, 321.3))
                 .docsRequestCategory(DocsCategory.CAFE.getCategory())
